requirements.txt
# Core Flask dependencies
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Cors==5.0.0
Flask-Session==0.5.0
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==23.0.0

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1

# API and Services
openai==1.59.7
stripe==11.5.0

# Utils
python-dotenv==1.0.0
email_validator==2.2.0

# Security
python-dateutil==2.9.0.post0
PyJWT==2.8.0

# Additional dependencies
Flask-Mail==0.9.1
